+++ packages/terminal-next/__tests__/browser/mock.service.ts
@@ -3,6 +3,7 @@ import { Terminal } from 'xterm';
 
 import { Injectable } from '@opensumi/di';
 import { WSChannel } from '@opensumi/ide-connection';
+import { createWSChannelForClient } from '@opensumi/ide-connection/__test__/common/ws-channel';
 import { WSWebSocketConnection } from '@opensumi/ide-connection/lib/common/connection';
 import { Disposable, PreferenceProvider, PreferenceResolveResult, PreferenceService } from '@opensumi/ide-core-browser';
 import { Deferred, Emitter, IDisposable, OperatingSystem, PreferenceScope, URI, uuid } from '@opensumi/ide-core-common';
@@ -69,7 +70,7 @@ export class MockTerminalService implements ITerminalService {
     launchConfig: IShellLaunchConfig,
   ): Promise<ITerminalConnection | undefined> {
     const sock = new WebSocket(localhost(getPort()));
-    const channel = WSChannel.forClient(new WSWebSocketConnection(sock), {
+    const channel = createWSChannelForClient(new WSWebSocketConnection(sock), {
       id: sessionId,
     });
